# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FramerAI REST API",
    "version": "1.0.0",
    "description": "OpenAPI 3.1 specification for the FramerAI backend REST API covering chat, generation, health, upload endpoints, and rate limits."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API server base path"
    }
  ],
  "paths": {
...
    "/generate/image": {
      "post": {
        "summary": "Generate image from text",
        "description": "Generates one or more images matching the prompt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "num_images": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 4,
                    "default": 1
                  },
                  "width": {
                    "type": "integer",
                    "minimum": 64,
                    "maximum": 2048
                  },
                  "height": {
                    "type": "integer",
                    "minimum": 64,
                    "maximum": 2048
                  },
                  "aspect": {
                    "type": "string",
                    "enum": [
                      "1:1",
                      "4:3",
                      "3:4",
                      "3:2",
                      "2:3",
                      "16:9",
                      "9:16",
                      "21:9"
                    ]
                  },
                  "tier": {
                    "type": "integer",
                    "enum": [
                      256,
                      512,
                      768,
                      1024
                    ]
                  },
                  "seed": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 2147483647
                  },
                  "resolution": {
                    "type": "integer",
                    "enum": [
                      64,
                      128,
                      256,
                      512
                    ],
                    "description": "Deprecated square-only resolution tier alias"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated image result",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
    "/generate/video": {
      "post": {
        "summary": "Generate video from text",
        "description": "Generates a video clip matching the prompt.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "num_frames": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 64,
                    "default": 16
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated video result",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
    "/generate/audio": {
      "post": {
        "summary": "Generate audio/speech from text",
        "description": "Generates audio or speech from text input.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated audio result",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
    "/generate/code": {
      "post": {
        "summary": "Generate code",
        "description": "Generates code snippets in the requested programming language.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "prompt"
                ],
                "properties": {
                  "prompt": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 4000
                  },
                  "language": {
                    "type": "string",
                    "enum": [
                      "python",
                      "javascript",
                      "typescript",
                      "java",
                      "go",
                      "rust",
                      "c",
                      "cpp",
                      "csharp",
                      "ruby",
                      "php",
                      "shell",
                      "sql",
                      "html",
                      "css"
                    ],
                    "default": "python"
                  },
                  "settings": {
                    "$ref": "#/components/schemas/GenerationSettings"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Generated code result",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeGenerationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
    "/generate/upload": {
      "post": {
        "summary": "Store a file for a later chat turn to reference",
        "description": "Stores an image or document and returns the path a chat message may attach. Runs no model: attaching is not asking.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Image or document to attach (image/*, application/pdf, text/plain, text/markdown, max file size 50MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stored attachment",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "path",
                    "kind"
                  ],
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "image",
                        "document",
                        "audio"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "size": {
                      "type": "integer"
                    },
                    "mimetype": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or unattachable file type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Uploaded file too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
    "/generate/document": {
      "post": {
        "summary": "Read an uploaded document",
        "description": "Uploads a document and returns its text in reading order, with page markers. Answers a prompt about the document when one is given. Pages that carry no text layer are reported in scannedPages rather than silently returned empty.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "document"
                ],
                "properties": {
                  "document": {
                    "type": "string",
                    "format": "binary",
                    "description": "Document to read (application/pdf, text/plain, text/markdown, max file size 50MB)"
                  },
                  "prompt": {
                    "type": "string",
                    "maxLength": 4000,
                    "description": "Optional question about the document."
                  },
                  "max_pages": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 2000,
                    "description": "Read at most this many pages."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Extracted document text",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "text",
                    "pages",
                    "documentPath"
                  ],
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "scannedPages": {
                      "type": "array",
                      "items": {
                        "type": "integer"
                      }
                    },
                    "content": {
                      "type": "string"
                    },
                    "documentPath": {
                      "type": "string"
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error, invalid file type, or unreadable document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "413": {
            "description": "Uploaded file too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "NOT_FOUND",
              "INVALID_JSON",
              "PAYLOAD_TOO_LARGE",
              "UPLOAD_ERROR",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "description": "Machine-readable error code"
          },
          "details": {
            "type": "array",
            "description": "Field validation error list (present for VALIDATION_ERROR or UPLOAD_ERROR)",
            "items": {
              "type": "object",
              "required": [
                "field",
                "message"
              ],
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ImageGenerationResponse": {
        "type": "object",
        "required": [
          "id",
          "prompt",
          "images",
          "metadata"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "prompt": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "placeholder"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "url": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "placeholder": {
                  "type": "boolean"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "VideoGenerationResponse": {
        "type": "object",
        "required": [
          "id",
          "prompt",
          "video",
          "metadata"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "prompt": {
            "type": "string"
          },
          "video": {
            "type": "object",
            "required": [
              "frames",
              "placeholder"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "frames": {
                "type": "integer"
              },
              "placeholder": {
                "type": "boolean"
              },
              "message": {
                "type": "string"
              }
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "AudioGenerationResponse": {
        "type": "object",
        "required": [
          "id",
          "prompt",
          "audio",
          "metadata"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "prompt": {
            "type": "string"
          },
          "audio": {
            "type": "object",
            "required": [
              "placeholder"
            ],
            "properties": {
              "url": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "placeholder": {
                "type": "boolean"
              },
              "message": {
                "type": "string"
              }
            }
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    },
    "headers": {
      "RateLimit-Limit": {
        "description": "Maximum requests allowed in rate-limit window",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimit-Remaining": {
        "description": "Remaining requests in current window",
        "schema": {
          "type": "integer"
        }
      },
      "RateLimit-Reset": {
        "description": "Seconds until rate limit window resets",
        "schema": {
          "type": "integer"
        }
      },
      "Retry-After": {
        "description": "Seconds to wait before retrying when rate-limited",
        "schema": {
          "type": "integer"
        }
      }
    }
  }
}